True 3-way merge using the shared android-mainline ancestry
(merge-base 54e301676a79 "Merge 5.4 into android-mainline").

5,866 conflicting paths resolved:
- 4,992 mechanical: our side only carried 5.4.y stable backports /
  ACK android11-5.4 patches already contained in 5.10.260 -> took 5.10.
- 874 vendor-touched: 5.10 base with CAF/Qualcomm/Xiaomi/KernelSU
  changes re-applied; CAF-internal stacks kept wholesale where the
  vendor driver ecosystem depends on them (qcom clk framework, SCM,
  UFS, sdhci-msm, dwc3, glink, rpmh/cmd-db, icc-rpmh, io-pgtable-arm,
  qcom-cpufreq-hw, qcom-pdc).

Kept device-critical downstream features: WALT scheduler hooks, techpack
build wiring + nested composite objects, lahaina/star DTS (vendor/qcom),
KernelSU hooks, ION/msm_dma_iommu, minidump, QGKI configs, Xiaomi
charger/thermal/touch/misc drivers, GKI vendor hooks unioned with ACK.

Vendor drivers are still 5.4-era in places and will need build-time API
fixups; see UPSTREAM_5.10_NOTES.md follow-ups.

+++ net/batman-adv/tvlv.h
@@ -1,5 +1,5 @@
 /* SPDX-License-Identifier: GPL-2.0 */
-/* Copyright (C) 2007-2019  B.A.T.M.A.N. contributors:
+/* Copyright (C) 2007-2020  B.A.T.M.A.N. contributors:
  *
  * Marek Lindner, Simon Wunderlich
  */
@@ -15,7 +15,7 @@
 void batadv_tvlv_container_register(struct batadv_priv *bat_priv,
 				    u8 type, u8 version,
 				    void *tvlv_value, u16 tvlv_value_len);
-u16 batadv_tvlv_container_ogm_append(struct batadv_priv *bat_priv,
+int batadv_tvlv_container_ogm_append(struct batadv_priv *bat_priv,
 				     unsigned char **packet_buff,
 				     int *packet_buff_len, int packet_min_len);
 void batadv_tvlv_ogm_receive(struct batadv_priv *bat_priv,
